// MesEquipment.Web/src/app/api/mes-equipment-api.ts
export interface CreateMachineDto {
  code: string;
  name: string;
  location: string;
  status: string;
}

export interface LoginDto {
  username: string;
  password: string;
}

export interface LoginResponseDto {
  token?: string | undefined;
  username?: string | undefined;
}

export interface MachineDto {
  id?: number;
  code?: string | undefined;
  name?: string | undefined;
  location?: string | undefined;
  status?: string | undefined;
  createdAt?: string;
}

export interface MachineDtoPagedResultDto {
  items?: MachineDto[] | undefined;
  totalCount?: number;
  page?: number;
  pageSize?: number;
}

export interface MachineStatusLogDto {
  id?: number;
  machineId?: number;
  oldStatus?: string | undefined;
  newStatus?: string | undefined;
  changedAt?: string;
  note?: string | undefined;
}

export interface UpdateMachineDto {
  code: string;
  name: string;
  location: string;
  status: string;
}